requirements.txt
Flask==2.0.3
requests==2.27.1
scikit-learn==1.0.2
numpy==1.21.5
tqdm==4.63.0
rich==12.0.0
seqeval==1.2.2
importlib-metadata==4.11.3
matplotlib==3.5.1
pandas==1.4.1
seaborn==0.11.2
torch==1.9.0
pytorch-crf==0.7.2

